Move data sourcing files to data_sources Fix bug related to avg_time
