+++ utils/data_sources/etherscan.py
@@ -1,7 +1,7 @@
 import requests
 import streamlit as st
 import pandas as pd
-from utils.data import average_time
+from utils.helpers import average_time
 
 ETHERSCAN_URL = 'https://api.etherscan.io/api'
 PK = st.secrets['etherscan_api']['KEY']
@@ -9,9 +9,10 @@ PK = st.secrets['etherscan_api']['KEY']
 def fetch_token_transfers(account):
     query = f'{ETHERSCAN_URL}?module=account&action=tokentx&address={account}&page=1&startblock=0&sort=asc&apikey={PK}'
     response = requests.get(query)
-    if response.status_code == 200:
+    if response.status_code == 200 and isinstance(response.json()['result'], list):
         df = pd.DataFrame(response.json()['result'])
-        return clean_token_transfers(df, account)
+        if len(df) > 0:
+            return clean_token_transfers(df, account)
     return pd.DataFrame()
 
 def clean_token_transfers(df, account):
@@ -30,7 +31,7 @@ def transfers_summary(df):
             "n_sent": len(df[df["direction"] == "sender"]),
             "n_received": len(df[df["direction"] == "receiver"]),
             "nunique_tokens": df["contractAddress"].nunique(),
-            "avg_time_transfers": average_time(df["date"]).round(2)
+            "avg_time_transfers": average_time(df["date"])
         }
     else:
         return {
